ENH: update pump to deal with missing edge
The harvard pump may start and stop faster than the polling frequency
so that epics missed the state change.  To work around this watch the
delivered volume and if it hits near the target, concluded that all status
objects are done

There is a motion status as a TTL on a serial port that should be connected
to EPICS as well, but this is a 'short-term' workaround.